Don't re-use the same slice each time
Re-using the same slice (probably my idea to reduce gc overhead) is a
bad idea because we pass that slice elsewhere where it is read
concurrently.  Ensure we don't write to if after we pass it along by
using a new slice each time.  Do this for both the query and the error
slices.