Auto (re-)connection
This adds a new BackgroundConnect option, which allows a caller of
Connect() to avoid thinking about database availabilty at startup and
while running.

This means the caller never sees connection errors at startup time if
the neo4j instance in unavailable. Instead the calling application can
start up, but all database operations will fail while there is no connection.

If an underlying error is detected while running that indicates a reconnection
is required, a new connection is created in the background.