Added Connect()
Added a Connect() function that should serve as the main entry point to
this package in future.  This deprecates various other types that are
currently exposed in favour of a single function with config options.
This will allow for greater flexibility in future once applcations are
all using this approach.  In particular, auto-reconnection and dealing
gracefully with an unavailable database will be possible.