Use rwapi.ConstraintOrTransactionError
Avoid returning errors that depend on neoism types. Instead use the
error in rwapi.  This helps minimise dependencies for applications that
depend on the rwapi API.  In particular, it means we can remove the
neoism dependency from our base application, making it suitable for
re-use with mongodb and in other contexts.

Our current consuming applications are already prepared to handle
rwapi.ConstraintOrTransactionError instead of
neoutils.ConstraintViolationError so this change does not require
further changes elsewhere.