Writable check addition.
Testing notes;

It's almost impossible to test the procedure `dbms.cluster.role()` since it's not supported when the tests are executed via circleci. The mode we run the neo4j in circleci is `SINGLE`, setting it to `CORE` enables the procedure but we need to start a full fledge neo4j cluster for this to happen.